Updated raspbian_jessie.md
Cleans up some english and spelling errors. Removes pronouns. Removed the word "can" in a lot of places. This is a tutorial that instructs. "Can" is more of a request instead of a specific instruction.

Just trying to help. No problem if you do not want to use this change